# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "訂餐系統",
        "title": "Order Food API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "/api/",
    "paths": {
...
        "/manager/changeshop": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更改商店",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "更改商店",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商店",
                        "name": "view",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "商店",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/manager/menu": {
            "get": {
                "description": "取得菜單",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shop"
                ],
                "summary": "取得菜單",
                "responses": {
                    "200": {
                        "description": "菜單",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}